+++ mfpdsp/builtin_phasor.c
@@ -94,19 +94,19 @@ config(mfp_processor * proc)
 
     /* get parameters */ 
     if (freq_ptr != NULL) {
-        d->const_freq = *(float *)freq_ptr;
+        d->const_freq = *(double *)freq_ptr;
         g_hash_table_remove(proc->params, "_sig_1");
         g_free(freq_ptr);
     }
 
     if (ampl_ptr != NULL) {
-        d->const_ampl = *(float *)ampl_ptr;
+        d->const_ampl = *(double *)ampl_ptr;
         g_hash_table_remove(proc->params, "_sig_2");
         g_free(ampl_ptr);
     }
 
     if (phase_ptr != NULL) {
-        d->phase = (double)(*(float *)phase_ptr);
+        d->phase = (double)(*(double *)phase_ptr);
         g_hash_table_remove(proc->params, "phase");
         g_free(phase_ptr);
     }
